[refactor]: di, dummy 폴더 위치 조정 (#35)
