+++ app/src/main/java/com/kuit/ourmenu/di/NetworkModule.kt
@@ -1,4 +1,4 @@
-package com.kuit.ourmenu.ui.dummy.di
+package com.kuit.ourmenu.di
 
 import com.google.gson.Gson
 import dagger.Module
