+++ app/src/main/java/com/kuit/ourmenu/di/ServiceModule.kt
@@ -1,6 +1,6 @@
-package com.kuit.ourmenu.ui.dummy.di
+package com.kuit.ourmenu.di
 
-import com.kuit.ourmenu.ui.dummy.service.DummyService
+import com.kuit.ourmenu.dummy.service.DummyService
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
