+++ app/src/main/java/com/kuit/ourmenu/dummy/service/DummyService.kt
@@ -1,0 +1,8 @@
+package com.kuit.ourmenu.dummy.service
+
+import com.kuit.ourmenu.dummy.data.DummyResponse
+
+interface DummyService {
+//    @GET("")
+    suspend fun getDummyData(): DummyResponse
+}
