--- app/src/main/java/com/kuit/ourmenu/ui/dummy/service/DummyService.kt
@@ -1,8 +1,0 @@
-package com.kuit.ourmenu.ui.dummy.service
-
-import com.kuit.ourmenu.ui.dummy.data.DummyResponse
-
-interface DummyService {
-//    @GET("")
-    suspend fun getDummyData(): DummyResponse
-}
